fix: email & password js 통합
