+++ signin.html
@@ -35,8 +35,7 @@
         </div>
       </div>
    </main>
-   <script type="module" src="/js/signin_email.js"></script>
-   <script type="module" src="/js/password.js"></script>
+   <script type="module" src="/js/signin.js"></script>
    <script type="module" src="/js/login.js"></script>
   </body>
 </html>
