+++ signup.html
@@ -42,8 +42,7 @@
       </div>
       </div>
     </main>
-    <script type="module" src="/js/signup_email.js"></script>
-    <script type="module" src="/js/password_check.js"></script>
+    <script type="module" src="/js/signup.js"></script>
     <script type="module" src="/js/login.js"></script>
   </body>
 </html>
